+++ src/main/kotlin/lotto/Application.kt
@@ -3,6 +3,7 @@ package lotto
 fun main() {
     // 의존성 주입
     val lottoView = LottoView()
-    val lottoService = LottoService()
-    LottoController(lottoView, lottoService).run()
+    val lottoGenerator = LottoGenerator()
+    val lottoRankFinder = LottoRankFinder()
+    LottoController(lottoView, lottoGenerator, lottoRankFinder).run()
 }
