+++ src/main/kotlin/lotto/LottoController.kt
@@ -5,8 +5,8 @@ import lotto.domain.LottoNumber
 
 class LottoController(
     private val lottoView: LottoView,
-    private val lottoService: LottoService,
     private val lottoGenerator: LottoGenerator,
+    private val lottoRankFinder: LottoRankFinder,
 ) {
     fun run() {
         // 순차적으로 실행...이게 과연 최선일까?
@@ -18,7 +18,7 @@ class LottoController(
         val winningLotto = Lotto(userInputWinningLotto.map { LottoNumber.of(it) })
         val bonusNum = lottoView.getBonusNum()
 
-        val rankMap = lottoService.getLottoRankMany(manyLotto, winningLotto, bonusNum)
+        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, winningLotto, bonusNum)
         lottoView.printResult(rankMap)
     }
 }
